requirements.txt
dash>=2.16.1
dash-mantine-components==0.14.1
dash-iconify>=0.1.2
plotly>=5.20.0
pandas>=2.2.0
faker>=24.0.0
gunicorn
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
APScheduler>=3.10.0
requests>=2.31.0
httpx
openpyxl>=3.1.0
fpdf2>=2.7.0
ldap3>=2.9.1
bcrypt>=4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
pydantic>=2.7.0